DocumentInformation rendering improvements: fetch info on tab switch
