+++ frontend/src/components/SourceDocument/DocumentInformation/Info/DocumentMetadataRow/DocumentMetadataGoToButton.tsx
