+++ frontend/src/components/SourceDocument/DocumentInformation/Info/DocumentMetadataRow/DocumentMetadataRow.tsx
@@ -2,19 +2,19 @@ import { ErrorMessage } from "@hookform/error-message";
 import { Stack } from "@mui/material";
 import { useCallback } from "react";
 import { SubmitErrorHandler, SubmitHandler, useForm } from "react-hook-form";
-import SdocMetadataHooks from "../../../../api/SdocMetadataHooks.ts";
+import SdocMetadataHooks from "../../../../../api/SdocMetadataHooks.ts";
 
-import { MetaType } from "../../../../api/openapi/models/MetaType.ts";
-import { SourceDocumentMetadataRead } from "../../../../api/openapi/models/SourceDocumentMetadataRead.ts";
-import { SourceDocumentMetadataReadResolved } from "../../../../api/openapi/models/SourceDocumentMetadataReadResolved.ts";
-import { SourceDocumentMetadataUpdate } from "../../../../api/openapi/models/SourceDocumentMetadataUpdate.ts";
-import { useOpenSnackbar } from "../../../../components/SnackbarDialog/useOpenSnackbar.ts";
-import { dateToLocaleYYYYMMDDString } from "../../../../utils/DateUtils.ts";
-import FormChipList from "../../../FormInputs/FormChipList.tsx";
-import FormDate from "../../../FormInputs/FormDate.tsx";
-import FormNumber from "../../../FormInputs/FormNumber.tsx";
-import FormSwitch from "../../../FormInputs/FormSwitch.tsx";
-import FormText from "../../../FormInputs/FormText.tsx";
+import { MetaType } from "../../../../../api/openapi/models/MetaType.ts";
+import { SourceDocumentMetadataRead } from "../../../../../api/openapi/models/SourceDocumentMetadataRead.ts";
+import { SourceDocumentMetadataReadResolved } from "../../../../../api/openapi/models/SourceDocumentMetadataReadResolved.ts";
+import { SourceDocumentMetadataUpdate } from "../../../../../api/openapi/models/SourceDocumentMetadataUpdate.ts";
+import { dateToLocaleYYYYMMDDString } from "../../../../../utils/DateUtils.ts";
+import FormChipList from "../../../../FormInputs/FormChipList.tsx";
+import FormDate from "../../../../FormInputs/FormDate.tsx";
+import FormNumber from "../../../../FormInputs/FormNumber.tsx";
+import FormSwitch from "../../../../FormInputs/FormSwitch.tsx";
+import FormText from "../../../../FormInputs/FormText.tsx";
+import { useOpenSnackbar } from "../../../../SnackbarDialog/useOpenSnackbar.ts";
 import MetadataEditMenu from "../MetadataEditMenu.tsx";
 import DocumentMetadataAddFilterButton from "./DocumentMetadataAddFilterButton.tsx";
 import DocumentMetadataGoToButton from "./DocumentMetadataGoToButton.tsx";
